unstructured/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
unstructured==0.12.5
pdfminer.six==20221105
pdf2image==1.16.3
markdownify==1.1.0
hanzidentifier==1.3.0
opencc-python-reimplemented==0.1.7
fastapi==0.110.0
uvicorn==0.29.0
starlette==0.36.3
tiktoken==0.9.0
redis==5.2.1